Refactor file_store models and tests into separate files (#2651)
